com.cubrid.cubridmigration.core.testfragment/src/com/cubrid/cubridmigration/core/engine/UserDefinedDataHandlerManagerTest.java
/*
 * Copyright (C) 2009 Search Solution Corporation. All rights reserved by Search Solution. 
 *
 * Redistribution and use in source and binary forms, with or without modification, 
 * are permitted provided that the following conditions are met: 
 *
 * - Redistributions of source code must retain the above copyright notice, 
 *   this list of conditions and the following disclaimer. 
 *
 * - Redistributions in binary form must reproduce the above copyright notice, 
 *   this list of conditions and the following disclaimer in the documentation 
 *   and/or other materials provided with the distribution. 
 *
 * - Neither the name of the <ORGANIZATION> nor the names of its contributors 
 *   may be used to endorse or promote products derived from this software without 
 *   specific prior written permission. 
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
 * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
 * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
 * IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
 * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
 * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, 
 * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
 * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
 * OF SUCH DAMAGE. 
 *
 */
package com.cubrid.cubridmigration.core.engine;

import java.util.HashMap;
import java.util.Map;

import junit.framework.Assert;

import org.junit.Test;

import com.cubrid.cubridmigration.core.common.PathUtils;

public class UserDefinedDataHandlerManagerTest {
	static {
		PathUtils.initPaths();
	}

	@Test
	public void test() {
		final UserDefinedDataHandlerManager instance = UserDefinedDataHandlerManager.getInstance();
		String dataHandler = "md5converter.jar:com.cubrid.migration.handler.MD5Converter";
		instance.putColumnDataHandler(dataHandler, true);
		instance.putColumnDataHandler(dataHandler, false);
		Object handler = instance.getColumnDataHandler(dataHandler);
		Assert.assertNotNull(handler);
		String columnName = "col1";
		Map<String, Object> recordMap = new HashMap<String, Object>();
		recordMap.put(columnName, "test");
		Object result = instance.handleColumnData(handler, recordMap, columnName);
		Assert.assertNotNull(result);
	}
}
